removed unused scripts
changed the color of paddle, made the gun invisible, and temporarily disabled ball split component from ball (so we can add it as powerup or gun mechanic).  Removed unused analytic scripts, we will be using firebase to collect data into a realtime database and can be viewable on a react app.